requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
httpx==0.26.0
pydantic==2.10.6
python-jose==3.3.0
python-multipart==0.0.9
aiosqlite==0.20.0
python-dotenv==1.0.1
pydantic-settings==2.8.1
greenlet==3.1.1 